Solve TypeError with Rankings by excepting TypeError
When the User reads a Book, and does not provide a ranking, the default ranking value was 'None'.

Fucntions were passing 'None' into a Book rating, which cause TypeErrors when methods and functions were expecting an int, but encountered None.

So I excepted TypeError  where possible, and things run smoothly.